django/requirements/base.txt
celery==5.0.5

Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-sendgrid-v5==0.9.1
django-taggit==1.3.0
django-taggit-labels==0.7.0
djangorestframework==3.12.2
drf-recaptcha==2.0.4
drf-yasg==1.20.0
python-decouple==3.4
python-twitter==3.5
redis==3.5.3
whitenoise==5.2.0